magic numbers added

+++ src/bit_wise/set_ithBit.java
@@ -2,7 +2,7 @@ package bit_wise;
 
 public class set_ithBit {
     public static void main(String[] args) {
-        int bit= 10001101;
+        int bit=  (byte) 10001101;
         //setting 5th bit as 1
         System.out.println(bit|(1<<4));
     }
